+++ src/process/validation/rules/__tests__/validateDate.test.ts
@@ -3,14 +3,14 @@ import { expect } from 'chai';
 import { FieldError } from 'error';
 import { calendarTextField, simpleDateTimeField, simpleTextField } from './fixtures/components';
 import { validateDate } from '../validateDate';
-import { generateProcessContext } from './fixtures/util';
+import { generateProcessorContext } from './fixtures/util';
 
 it('Validating a component without a date/time concern will return null', async () => {
     const component = simpleTextField;
     const data = {
         component: 'Hello, world!',
     };
-    const context = generateProcessContext(component, data);
+    const context = generateProcessorContext(component, data);
     const result = await validateDate(context);
     expect(result).to.equal(null);
 });
@@ -18,7 +18,7 @@ it('Validating a component without a date/time concern will return null', async 
 it('Validating a date/time component with no data will return null', async () => {
     const component = simpleDateTimeField;
     const data = {};
-    const context = generateProcessContext(component, data);
+    const context = generateProcessorContext(component, data);
     const result = await validateDate(context);
     expect(result).to.equal(null);
 });
@@ -28,7 +28,7 @@ it('Validating a date/time component with an invalid date string value will retu
     const data = {
         component: 'hello, world!',
     };
-    const context = generateProcessContext(component, data);
+    const context = generateProcessorContext(component, data);
     const result = await validateDate(context);
     expect(result).to.be.instanceOf(FieldError);
     expect(result?.errorKeyOrMessage).to.equal('invalidDate');
@@ -39,7 +39,7 @@ it('Validating a date/time component with an valid date string value will return
     const data = {
         component: '2023-03-09T12:00:00-06:00',
     };
-    const context = generateProcessContext(component, data);
+    const context = generateProcessorContext(component, data);
     const result = await validateDate(context);
     expect(result).to.equal(null);
 });
@@ -49,7 +49,7 @@ it('Validating a date/time component with an invalid Date object will return a F
     const data = {
         component: new Date('Hello, world!'),
     };
-    const context = generateProcessContext(component, data);
+    const context = generateProcessorContext(component, data);
     const result = await validateDate(context);
     expect(result).to.be.instanceOf(FieldError);
     expect(result?.errorKeyOrMessage).to.equal('invalidDate');
@@ -60,7 +60,7 @@ it('Validating a date/time component with a valid Date object will return null',
     const data = {
         component: new Date(),
     };
-    const context = generateProcessContext(component, data);
+    const context = generateProcessorContext(component, data);
     const result = await validateDate(context);
     expect(result).to.equal(null);
 });
@@ -68,7 +68,7 @@ it('Validating a date/time component with a valid Date object will return null',
 it('Validating a textField calendar picker component with no data will return null', async () => {
     const component = calendarTextField;
     const data = {};
-    const context = generateProcessContext(component, data);
+    const context = generateProcessorContext(component, data);
     const result = await validateDate(context);
     expect(result).to.equal(null);
 });
@@ -78,7 +78,7 @@ it('Validating a textField calendar picker component with an invalid date string
     const data = {
         component: 'hello, world!',
     };
-    const context = generateProcessContext(component, data);
+    const context = generateProcessorContext(component, data);
     const result = await validateDate(context);
     expect(result).to.be.instanceOf(FieldError);
     expect(result?.errorKeyOrMessage).to.equal('invalidDate');
@@ -89,7 +89,7 @@ it('Validating a textField calendar picker component with an valid date string v
     const data = {
         component: '2023-03-09T12:00:00-06:00',
     };
-    const context = generateProcessContext(component, data);
+    const context = generateProcessorContext(component, data);
     const result = await validateDate(context);
     expect(result).to.equal(null);
 });
@@ -99,7 +99,7 @@ it('Validating a textField calendar picker component with an invalid Date object
     const data = {
         component: new Date('Hello, world!'),
     };
-    const context = generateProcessContext(component, data);
+    const context = generateProcessorContext(component, data);
     const result = await validateDate(context);
     expect(result).to.be.instanceOf(FieldError);
     expect(result?.errorKeyOrMessage).to.equal('invalidDate');
@@ -110,7 +110,7 @@ it('Validating a textField calendar picker component with a valid Date object wi
     const data = {
         component: new Date(),
     };
-    const context = generateProcessContext(component, data);
+    const context = generateProcessorContext(component, data);
     const result = await validateDate(context);
     expect(result).to.equal(null);
 });