+++ src/process/validation/rules/__tests__/validateDay.test.ts
@@ -2,7 +2,7 @@ import { expect } from 'chai';
 
 import { FieldError } from 'error';
 import { simpleDayField, simpleTextField } from './fixtures/components';
-import { generateProcessContext } from './fixtures/util';
+import { generateProcessorContext } from './fixtures/util';
 import { validateDay } from '../validateDay';
 
 it('Validating a non-day component will return null', async () => {
@@ -10,7 +10,7 @@ it('Validating a non-day component will return null', async () => {
     const data = {
         component: 'Hello, world!',
     };
-    const context = generateProcessContext(component, data);
+    const context = generateProcessorContext(component, data);
     const result = await validateDay(context);
     expect(result).to.equal(null);
 });
@@ -20,7 +20,7 @@ it('Validating a day component with an invalid date string value will return a F
     const data = {
         component: 'hello, world!',
     };
-    const context = generateProcessContext(component, data);
+    const context = generateProcessorContext(component, data);
     const result = await validateDay(context);
     expect(result).to.be.instanceOf(FieldError);
     expect(result?.errorKeyOrMessage).to.equal('invalidDay');
@@ -31,7 +31,7 @@ it('Validating a day component with an valid date string value will return null'
     const data = {
         component: '03/23/2023',
     };
-    const context = generateProcessContext(component, data);
+    const context = generateProcessorContext(component, data);
     const result = await validateDay(context);
     expect(result).to.equal(null);
 });
@@ -41,7 +41,7 @@ it('Validating a day component with an invalid Date object will return a FieldEr
     const data = {
         component: new Date('Hello, world!'),
     };
-    const context = generateProcessContext(component, data);
+    const context = generateProcessorContext(component, data);
     const result = await validateDay(context);
     expect(result).to.be.instanceOf(FieldError);
     expect(result?.errorKeyOrMessage).to.equal('invalidDay');
@@ -52,7 +52,7 @@ it('Validating a day component with a valid Date object will return a field erro
     const data = {
         component: new Date(),
     };
-    const context = generateProcessContext(component, data);
+    const context = generateProcessorContext(component, data);
     const result = await validateDay(context);
     expect(result).to.be.instanceOf(FieldError);
     expect(result?.errorKeyOrMessage).to.equal('invalidDay');