+++ src/process/validation/rules/__tests__/validateMinimumYear.test.ts
@@ -3,7 +3,7 @@ import { expect } from 'chai';
 import { DayComponent } from 'types';
 import { FieldError } from 'error';
 import { simpleDayField, simpleTextField } from './fixtures/components';
-import { generateProcessContext } from './fixtures/util';
+import { generateProcessorContext } from './fixtures/util';
 import { validateMinimumYear } from '../validateMinimumYear';
 
 it('Validating a component without the minYear parameter will return null', async () => {
@@ -11,7 +11,7 @@ it('Validating a component without the minYear parameter will return null', asyn
     const data = {
         component: 'Hello, world!',
     }
-    const context = generateProcessContext(component, data);
+    const context = generateProcessorContext(component, data);
     const result = await validateMinimumYear(context);
     expect(result).to.equal(null);
 });
@@ -21,7 +21,7 @@ it('Validating a day component without the minYear parameter will return null', 
     const data = {
         component: '01/22/2023',
     };
-    const context = generateProcessContext(component, data);
+    const context = generateProcessorContext(component, data);
     const result = await validateMinimumYear(context);
     expect(result).to.equal(null);
 });
@@ -38,7 +38,7 @@ it('Validating a day component with the minYear parameter will return a FieldErr
     const data = {
         component: '01/22/2022',
     };
-    const context = generateProcessContext(component, data);
+    const context = generateProcessorContext(component, data);
     const result = await validateMinimumYear(context);
     expect(result).to.be.instanceOf(FieldError);
     expect(result?.errorKeyOrMessage).to.contain('minYear');
@@ -56,7 +56,7 @@ it('Validating a day component with the minYear parameter will return null if th
     const data = {
         component: '01/22/2022',
     };
-    const context = generateProcessContext(component, data);
+    const context = generateProcessorContext(component, data);
     const result = await validateMinimumYear(context);
     expect(result).to.equal(null);
 });
@@ -73,7 +73,7 @@ it('Validating a day component with the minYear parameter will return null if th
     const data = {
         component: '01/22/2023',
     };
-    const context = generateProcessContext(component, data);
+    const context = generateProcessorContext(component, data);
     const result = await validateMinimumYear(context);
     expect(result).to.equal(null);
 });